PhoneAssistant.Tests/Features/AddItem/AddItemViewModelTests.cs
using Moq.AutoMock;

using Xunit;
using PhoneAssistant.WPF.Application.Entities;
using PhoneAssistant.WPF.Features.AddItem;
using PhoneAssistant.WPF.Application.Repositories;
using Moq;
using System.ComponentModel.DataAnnotations;
using PhoneAssistant.WPF.Application;
using CommunityToolkit.Mvvm.Messaging;
using PhoneAssistant.Tests.Shared;
using System.ComponentModel;

namespace PhoneAssistant.Tests.Features.AddItem;
public class AddItemViewModelTests
{
    private readonly AutoMocker _mocker = new AutoMocker();
    private readonly AddItemViewModel _sut;

    public AddItemViewModelTests()
    {
        _sut = _mocker.CreateInstance<AddItemViewModel>();
    }

    #region Phone
    [Fact]
    void CanSavePhone_ShouldBeEnabled_WhenAllRequiredPropertiesSupplied()
    {
        Mock<IPhonesRepository> _repository = _mocker.GetMock<IPhonesRepository>();
        _repository.Setup(r => r.AssetTagUniqueAsync("MP00001")).ReturnsAsync(true);

        _sut.AssetTag = "MP00001";
        _sut.Condition = "condition";
        _sut.Imei = "355808981147090";
        _sut.Model = "model";
        _sut.Status = "status";
        
        _mocker.VerifyAll();
        Assert.True(_sut.CanSavePhone());
    }

    [Fact]
    void GetErrors_ShouldContainAssetTagError_WhenAssetTagNotUnique()
    {
        Mock<IPhonesRepository> _repository = _mocker.GetMock<IPhonesRepository>();
        _repository.Setup(r => r.AssetTagUniqueAsync("MP99999")).ReturnsAsync(false);

        _sut.AssetTag = "MP99999";

        _mocker.VerifyAll();
        IEnumerable<ValidationResult> errors = _sut.GetErrors(nameof(_sut.AssetTag));
        Assert.Equal("Asset Tag must be unique", errors.First().ToString());
    }

    [Theory]
    [InlineData("Decommissioned")]
    [InlineData("Disposed")]
    public void GetErrors_ShouldContainError_WhenDisposalWithDefaultTicket(string status)
    {
        Mock<IUserSettings> setings = _mocker.GetMock<IUserSettings>();
        setings.Setup(s => s.DefaultDecommissionedTicket).Returns(123456);

        _sut.Status = status;

        IEnumerable<ValidationResult> errors = _sut.GetErrors(nameof(_sut.Ticket));
        Assert.Empty(errors);
    }

    [Theory]
    [InlineData("Decommissioned")]
    [InlineData("Disposed")]
    public void GetErrors_ShouldContainError_WhenDisposalWithNoDefaultTicket(string status)
    {
        _sut.Status = status;

        IEnumerable<ValidationResult> errors = _sut.GetErrors(nameof(_sut.Ticket));
        Assert.Equal("Ticket must 6 or 7 digits", errors.First().ToString());
    }

    [Theory]
    [InlineData(" ")]
    [InlineData("12345")]
    [InlineData("12345678")]
    [InlineData("1A345")]
    void GetErrors_ShouldContainTicketError_WhenTicketInvalid(string ticket)
    {
        Mock<IPhonesRepository> _repository = _mocker.GetMock<IPhonesRepository>();

        _sut.Ticket = ticket;

        IEnumerable<ValidationResult> errors = _sut.GetErrors(nameof(_sut.Ticket));
        Assert.Equal("Ticket must 6 or 7 digits", errors.First().ToString());
    }

    [Theory]
    [InlineData("PC00001")]
    [InlineData("MP00002")]
    void GetErrors_ShouldBeEmpty_WhenAssetTagFormatValid(string assetTag)
    {
        Mock<IPhonesRepository> _repository = _mocker.GetMock<IPhonesRepository>();
        _repository.Setup(r => r.AssetTagUniqueAsync(assetTag)).ReturnsAsync(true);

        _sut.AssetTag = assetTag;

        _mocker.VerifyAll();
        IEnumerable<ValidationResult> errors = _sut.GetErrors(nameof(_sut.AssetTag));
        Assert.Empty(errors);
    }

    [Fact]
    void GetErrors_ShouldBeEmpty_WhenAssetTagUnique()
    {
        Mock<IPhonesRepository> _repository = _mocker.GetMock<IPhonesRepository>();
        _repository.Setup(r => r.AssetTagUniqueAsync("MP99999")).ReturnsAsync(true);

        _sut.AssetTag = "MP99999";

        _mocker.VerifyAll();
        IEnumerable<ValidationResult> errors = _sut.GetErrors(nameof(_sut.AssetTag));
        Assert.Empty(errors);
    }

    [Fact]
    void GetErrors_ShouldBeEmpty_WhenIMEISet()
    {
        _sut.Imei = "355808981147090";

        IEnumerable<ValidationResult> errors = _sut.GetErrors(nameof(_sut.Imei));
        Assert.Empty(errors);
    }

    [Fact]
    void GetErrors_ShouldBeEmpty_WhenModelSet()
    {
        _sut.Model = "model";

        IEnumerable<ValidationResult> errors = _sut.GetErrors(nameof(_sut.Model));
        Assert.Empty(errors);
    }

    [Fact]
    void PhoneClearCommand_ShouldDisablePhoneSave()
    {
        _sut.PhoneClearCommand.Execute(null);

        Assert.False(_sut.CanSavePhone());
    }

    [Fact]
    void PhoneClearCommand_ShouldResetAllProperties()
    {
        ArrangeSetAllPhoneProperties();

        _sut.PhoneClearCommand.Execute(null);

        AssertResetAllPhoneProperties();
    }

    private void ArrangeSetAllPhoneProperties()
    {
        _sut.AssetTag = "MP00000";
        _sut.Condition = "condition";
        _sut.FormerUser = "former user";
        _sut.Imei = "imei";
        _sut.Model = "model";
        _sut.PhoneNotes = "notes";
        _sut.OEM = OEMs.Samsung;
        _sut.Status = "status";
        _sut.Ticket = 7654321.ToString();
    }

    private void AssertResetAllPhoneProperties()
    {
        Assert.Null(_sut.AssetTag);
        Assert.Equal(ApplicationSettings.Conditions[1].Substring(0, 1), _sut.Condition);
        Assert.Null( _sut.FormerUser);
        Assert.Equal(string.Empty, _sut.Imei);
        Assert.Equal(string.Empty, _sut.Model);
        Assert.Null(_sut.PhoneNotes);
        Assert.Equal(OEMs.Samsung,_sut.OEM);
        Assert.Equal(ApplicationSettings.Statuses[1], _sut.Status);
        Assert.Null(_sut.Ticket);
    }

    [Fact]
    [Description("Issue 53")]
    void PhoneSaveCommand_ShouldNotSaveSimDetails()
    {
        Mock<IPhonesRepository> repository = _mocker.GetMock<IPhonesRepository>();
        repository.Setup(r => r.AssetTagUniqueAsync("MP00001")).ReturnsAsync(true);        
        Phone actual = new() { Condition = "", Imei = "", Model = "", OEM = OEMs.Apple, Status = "" };
        repository.Setup(r => r.CreateAsync(It.IsAny<Phone>())).Callback<Phone>(p => actual = p);

        _sut.AssetTag = "MP00001";
        _sut.Imei = "355808981147090";
        _sut.Model = "model";
        _sut.PhoneNumber = "07123456789";
        _sut.SimNumber = "8944122605566849402";

        _sut.PhoneSaveCommand.Execute(null);

        _mocker.VerifyAll();
        Assert.Null(actual.PhoneNumber);
        Assert.Null(actual.SimNumber);
    }


    [Fact]
    void PhoneSaveCommand_ShouldCallRepository()
    {
        Mock<IPhonesRepository> _repository = _mocker.GetMock<IPhonesRepository>();
        _repository.Setup(r => r.AssetTagUniqueAsync("MP00001")).ReturnsAsync(true);
        _repository.Setup(r => r.CreateAsync(It.IsAny<Phone>()));

        _sut.AssetTag = "MP00001";
        _sut.Imei = "355808981147090";
        _sut.Model = "model";

        _sut.PhoneSaveCommand.Execute(null);

        _mocker.VerifyAll();
    }

    [Fact]
    void PhoneSaveCommand_ShouldDisablePhoneSave()
    {
        _sut.PhoneSaveCommand.Execute(null);

        Assert.False(_sut.CanSavePhone());
    }

    [Fact]
    void PhoneSaveCommand_ShouldResetAllPhoneProperties()
    {
        ArrangeSetAllPhoneProperties();

        _sut.PhoneSaveCommand.Execute(null);

        AssertResetAllPhoneProperties();
    }

    [Fact]
    void PhoneSaveCommand_ShouldSendPhoneMessage()
    {
        Mock<IMessenger> message = _mocker.GetMock<IMessenger>();
        message.Setup(m => m.Send(It.IsAny<Phone>(), It.IsAny<IsAnyToken>()));

        _sut.AssetTag = "MP00001";
        _sut.Imei = "355808981147090";
        _sut.Model = "model";

        _sut.PhoneSaveCommand.Execute(null);

        message.Verify(x => x.Send(It.IsAny<Phone>(), It.IsAny<IsAnyToken>()), Times.Once);
    }

    [Fact]
    void ValidateImei_ShouldReturnError_WhenIMEIEmptyOrWhiteSpace()
    {
        ValidationContext ctx = new(_sut, null, null);

        ValidationResult actual1 = AddItemViewModel.ValidateImeiAsync("", ctx);
        ValidationResult actual2 = AddItemViewModel.ValidateImeiAsync("  ", ctx);

        Assert.NotNull(actual1);
        Assert.Equal("IMEI is required", actual1.ErrorMessage);
        Assert.NotNull(actual2);
        Assert.Equal("IMEI is required", actual2.ErrorMessage);
    }

    [Fact]
    void ValidateImei_ShouldReturnError_WhenIMEINotNumeric()
    {
        ValidationContext ctx = new(_sut, null, null);

        ValidationResult actual = AddItemViewModel.ValidateImeiAsync("abc", ctx);

        Assert.NotNull(actual);
        Assert.Equal("IMEI check digit incorrect", actual.ErrorMessage);
    }

    [Fact]
    void ValidateImei_ShouldReturnError_WhenIMEINotUnique()
    {
        ValidationContext ctx = new(_sut, null, null);
        Mock<IPhonesRepository> _repository = new Mock<IPhonesRepository>();
        _repository = _mocker.GetMock<IPhonesRepository>();
        _repository.Setup(p => p.ExistsAsync("353427866717729")).ReturnsAsync(true);

        ValidationResult actual = AddItemViewModel.ValidateImeiAsync("353427866717729", ctx);
        _repository.VerifyAll();
        Assert.NotNull(actual);
        Assert.Equal("IMEI must be unique", actual.ErrorMessage);
    }

    [Fact]
    void ValidateIMEI_ShouldReturnError_WhenIMEIInvalid()
    {
        ValidationContext ctx = new(_sut, null, null);

        ValidationResult actual = AddItemViewModel.ValidateImeiAsync("355808981132899", ctx); // An invalid 15-digit IMEI

        Assert.NotNull(actual);
        Assert.Equal("IMEI check digit incorrect", actual.ErrorMessage);
    }

    [Fact]
    void ValidateIMEI_ShouldReturnValidResult_WhenIMEIValid()
    {
        ValidationContext ctx = new(_sut, null, null);

        ValidationResult actual = AddItemViewModel.ValidateImeiAsync("355808981132845", ctx); // A valid 15-digit IMEI

        Assert.Equal(ValidationResult.Success, actual);
    }
    #endregion

    #region SIM
    [Fact]
    void CanDeleteSIM_ShouldBeEnabled_WhenAllRequiredPropertiesSuppliedAndSIMExists()
    {
        Mock<IPhonesRepository> phones = new Mock<IPhonesRepository>();
        phones = _mocker.GetMock<IPhonesRepository>();
        phones.Setup(p => p.PhoneNumberExistsAsync("07123456789")).ReturnsAsync(false);

        Mock<ISimsRepository> sims = _mocker.GetMock<ISimsRepository>();
        sims.Setup(r => r.GetSIMNumberAsync("07123456789")).ReturnsAsync("8944122605566849402");

        _sut.PhoneNumber = "07123456789";

        _mocker.VerifyAll();
        IEnumerable<ValidationResult> errors = _sut.GetErrors(nameof(_sut.PhoneNumber));
        Assert.True(_sut.CanDeleteSIM());
    }

    [Fact]
    void CanDeleteSIM_ShouldBeDisabled_WhenSIMDoesNotExist()
    {
        Mock<ISimsRepository> _repository = _mocker.GetMock<ISimsRepository>();
#pragma warning disable CS8600 // Converting null literal or possible null value to non-nullable type.
        _repository.Setup(r => r.GetSIMNumberAsync("07123456789")).ReturnsAsync((string)null);
#pragma warning restore CS8600 // Converting null literal or possible null value to non-nullable type.

        _sut.PhoneNumber = "07123456789";
        _sut.SimNumber = "8944122605566849402";

        _mocker.VerifyAll();
        Assert.False(_sut.CanDeleteSIM());
    }

    [Fact]
    void CanSaveSIM_ShouldBeEnabled_WhenAllRequiredPropertiesSupplied()
    {
        Mock<ISimsRepository> _repository = _mocker.GetMock<ISimsRepository>();
#pragma warning disable CS8600 // Converting null literal or possible null value to non-nullable type.
        _repository.Setup(r => r.GetSIMNumberAsync("07123456789")).ReturnsAsync((string)null);
#pragma warning restore CS8600 // Converting null literal or possible null value to non-nullable type.

        _sut.PhoneNumber = "07123456789";
        _sut.SimNumber = "8944122605566849402";

        _mocker.VerifyAll();
        Assert.True(_sut.CanSaveSIM());
    }

    [Fact]
    void GetErrors_ShouldBeEmpty_WhenPhoneNumberSet()
    {
        _sut.PhoneNumber = "07123456789";

        IEnumerable<ValidationResult> errors = _sut.GetErrors(nameof(_sut.PhoneNumber));
        Assert.Empty(errors);
    }

    [Fact]
    void GetErrors_ShouldBeEmpty_WhenSimNumberSet()
    {
        _sut.SimNumber = "8944122605566849402";

        IEnumerable<ValidationResult> errors = _sut.GetErrors(nameof(_sut.SimNumber));
        Assert.Empty(errors);
    }

    [Fact]
    void OnPhoneNumberChanged_ShouldSetSimNumber_WhenSimExists()
    {
        Mock<ISimsRepository> _repository = _mocker.GetMock<ISimsRepository>();
        _repository.Setup(r => r.GetSIMNumberAsync("07123456789")).ReturnsAsync("sim number");

        _sut.PhoneNumber = "07123456789";

        _mocker.VerifyAll();
        Assert.Equal("sim number", _sut.SimNumber);
        Assert.False(_sut.CanSaveSIM());
    }

    [Fact]
    void SIMClearCommand_ShouldDisableSIMSave()
    {
        _sut.SIMClearCommand.Execute(null);

        Assert.False(_sut.CanSaveSIM());
    }

    [Fact]
    void SIMClearCommand_ShouldResetAllSIMProperties()
    {
        _sut.PhoneNumber = "1234567890";
        _sut.SimNumber = "1234567890123456789";
        _sut.SimNotes = "Notes";

        _sut.SIMClearCommand.Execute(null);

        Assert.Null(_sut.PhoneNumber);
        Assert.Null(_sut.SimNumber);
        Assert.Null(_sut.SimNotes);
    }

    [Fact]
    void SimDeleteCommand_ShouldCallRepository()
    {
        Mock<ISimsRepository> _repository = _mocker.GetMock<ISimsRepository>();
        _repository.Setup(r => r.GetSIMNumberAsync("07123456789")).ReturnsAsync("8944122605566849402");
        _repository.Setup(r => r.DeleteSIMAsync("07123456789")).ReturnsAsync("8944122605566849402");
        _sut.PhoneNumber = "07123456789";

        _sut.SIMDeleteCommand.Execute(null);

        _mocker.VerifyAll();
    }

    [Fact]
    void SimSaveCommand_ShouldCallRepository()
    {
        Sim newSIM = new() { PhoneNumber = "", SimNumber = "" };
        Mock<ISimsRepository> _repository = _mocker.GetMock<ISimsRepository>();
#pragma warning disable CS8600 // Converting null literal or possible null value to non-nullable type.
        _repository.Setup(r => r.GetSIMNumberAsync("07123456789")).ReturnsAsync((string)null);
#pragma warning restore CS8600 // Converting null literal or possible null value to non-nullable type.
        _repository.Setup(r => r.CreateAsync(It.IsAny<Sim>())).Callback<Sim>(s => newSIM = s);
        _sut.PhoneNumber = "07123456789";
        _sut.SimNumber = "8944122605566849402";

        _sut.SIMSaveCommand.Execute(null);

        Assert.Equal("07123456789", newSIM.PhoneNumber);
        Assert.Equal("8944122605566849402", newSIM.SimNumber);

        _mocker.VerifyAll();
    }

    [Fact]
    void SimSaveCommand_ShouldDisablePhoneSave()
    {
        _sut.SIMSaveCommand.Execute(null);

        Assert.False(_sut.CanSaveSIM());
    }

    [Fact]
    void SimSaveCommand_ShouldResetAllPhoneProperties()
    {
        _sut.PhoneNumber = "1234567890";
        _sut.SimNumber = "1234567890123456789";
        _sut.SimNotes = "Notes";

        _sut.SIMSaveCommand.Execute(null);

        Assert.Null(_sut.PhoneNumber);
        Assert.Null(_sut.SimNumber);
        Assert.Null(_sut.SimNotes);
    }

    [Fact]
    void SimSaveCommand_ShouldSendSimMessage()
    {
        Mock<IMessenger> message = _mocker.GetMock<IMessenger>();
        message.Setup(m => m.Send(It.IsAny<Sim>(), It.IsAny<IsAnyToken>()));
        
        _sut.PhoneNumber = "1234567890";
        _sut.SimNumber = "1234567890123456789";

        _sut.SIMSaveCommand.Execute(null);

        message.Verify(x => x.Send(It.IsAny<Sim>(), It.IsAny<IsAnyToken>()), Times.Once);
    }

    [Fact]
    void ValidateSimNumber_ShouldReturnError_WhenSimNumberEmptyOrWhiteSpace()
    {
        ValidationContext ctx = new(_sut, null, null);

        ValidationResult actual1 = AddItemViewModel.ValidateSimNumber("", ctx);
        ValidationResult actual2 = AddItemViewModel.ValidateSimNumber("  ", ctx);

        Assert.NotNull(actual1);
        Assert.Equal("SIM Number is required", actual1.ErrorMessage);
        Assert.NotNull(actual2);
        Assert.Equal("SIM Number is required", actual2.ErrorMessage);
    }

    [Fact]
    void ValidateSimNumber_ShouldReturnError_WhenSimNumberNotNumeric()
    {
        ValidationContext ctx = new(_sut, null, null);

        ValidationResult actual = AddItemViewModel.ValidateSimNumber("abc", ctx);

        Assert.NotNull(actual);
        Assert.Equal("SIM Number check digit incorrect", actual.ErrorMessage);
    }

    [Fact]
    void ValidateSimNumber_ShouldReturnError_WhenSimNumberInvalid()
    {
        ValidationContext ctx = new(_sut, null, null);

        ValidationResult actual = AddItemViewModel.ValidateSimNumber("8944125605540324744", ctx); // An invalid 15-digit SIM Number

        Assert.NotNull(actual);
        Assert.Equal("SIM Number check digit incorrect", actual.ErrorMessage);
    }

    [Fact]
    void ValidateSimNumber_ShouldReturnValidResult_WhenSimNumberValid()
    {
        ValidationContext ctx = new(_sut, null, null);

        ValidationResult actual = AddItemViewModel.ValidateSimNumber("8944125605540324743", ctx); // A valid 19-digit SimNumber

        Assert.Equal(ValidationResult.Success, actual);
    }
    #endregion

    #region PhoneWithSim
    [Fact]
    void CanSavePhoneWithSIM_ShouldBeEnabled_WhenAllRequiredPropertiesSupplied()
    {
        Mock<IPhonesRepository> _phones = _mocker.GetMock<IPhonesRepository>();
        _phones.Setup(r => r.AssetTagUniqueAsync("MP00001")).ReturnsAsync(true);
        Mock<ISimsRepository> sims = _mocker.GetMock<ISimsRepository>();
#pragma warning disable CS8600 // Converting null literal or possible null value to non-nullable type.
        sims.Setup(r => r.GetSIMNumberAsync("07123456789")).ReturnsAsync((string)null);
#pragma warning restore CS8600 // Converting null literal or possible null value to non-nullable type.

        _sut.AssetTag = "MP00001";
        _sut.Condition = "condition";
        _sut.Imei = "355808981147090";
        _sut.Model = "model";
        _sut.Status = "status";
        _sut.PhoneNumber = "07123456789";
        _sut.SimNumber = "8944122605566849402";

        _mocker.VerifyAll();
        Assert.True(_sut.CanSavePhoneWithSIM());
    }

    [Fact]
    void PhoneWithSIMCommand_ShouldCallRepository()
    {
        Phone actual = new() { Condition = "", Imei = "", Model = "", OEM = OEMs.Apple, Status = "" };
        Mock<IPhonesRepository> _phones = _mocker.GetMock<IPhonesRepository>();
        _phones.Setup(r => r.AssetTagUniqueAsync("MP00001")).ReturnsAsync(true);
        _phones.Setup(r => r.CreateAsync(It.IsAny<Phone>())).Callback<Phone>(p => actual = p); 
        Mock<ISimsRepository> sims = _mocker.GetMock<ISimsRepository>();
#pragma warning disable CS8600 // Converting null literal or possible null value to non-nullable type.
        sims.Setup(r => r.GetSIMNumberAsync("07123456789")).ReturnsAsync((string)null);
#pragma warning restore CS8600 // Converting null literal or possible null value to non-nullable type.

        _sut.AssetTag = "MP00001";
        _sut.Condition = "condition";
        _sut.Imei = "355808981147090";
        _sut.Model = "model";
        _sut.Status = "status";
        _sut.PhoneNumber = "07123456789";
        _sut.SimNumber = "8944122605566849402";

        _sut.PhoneWithSIMSaveCommand.Execute(null);

        _mocker.VerifyAll();
        Assert.Equal("07123456789", actual.PhoneNumber);
        Assert.Equal("8944122605566849402", actual.SimNumber);

    }

    [Fact]
    void PhoneWithSIMCommand_ShouldDeleteSIM_WhenSimExists()
    {
        Mock<IPhonesRepository> _phones = _mocker.GetMock<IPhonesRepository>();
        _phones.Setup(r => r.AssetTagUniqueAsync("MP00001")).ReturnsAsync(true);
        _phones.Setup(r => r.CreateAsync(It.IsAny<Phone>()));
        Mock<ISimsRepository> sims = _mocker.GetMock<ISimsRepository>();
        sims.Setup(r => r.GetSIMNumberAsync("07123456789")).ReturnsAsync("8944122605566849402");
        sims.Setup(s => s.DeleteSIMAsync("07123456789")).ReturnsAsync("8944122605566849402");

        _sut.AssetTag = "MP00001";
        _sut.Condition = "condition";
        _sut.Imei = "355808981147090";
        _sut.Model = "model";
        _sut.Status = "status";
        _sut.PhoneNumber = "07123456789";

        _sut.PhoneWithSIMSaveCommand.Execute(null);

        _mocker.VerifyAll();
    }
    [Fact]
    void ValidatePhoneNumber_ShouldReturnError_WhenPhoneNumberNotUnique()
    {
        ValidationContext ctx = new(_sut, null, null);
        Mock<IPhonesRepository> _repository = new Mock<IPhonesRepository>();
        _repository = _mocker.GetMock<IPhonesRepository>();
        _repository.Setup(p => p.PhoneNumberExistsAsync("07123456789")).ReturnsAsync(true);
        
        ValidationResult actual = AddItemViewModel.ValidatePhoneNumber("07123456789", ctx);

        _repository.VerifyAll();
        Assert.NotNull(actual);
        Assert.Equal("Phone Number already linked to phone", actual.ErrorMessage);
    }

    [Fact]
    void ValidatePhoneNumber_ShouldReturnValidResult_WhenPhoneNumberUnique()
    {
        ValidationContext ctx = new(_sut, null, null);
        Mock<IPhonesRepository> _repository = new Mock<IPhonesRepository>();
        _repository = _mocker.GetMock<IPhonesRepository>();
        _repository.Setup(p => p.PhoneNumberExistsAsync("07123456789")).ReturnsAsync(false);
        
        ValidationResult actual = AddItemViewModel.ValidatePhoneNumber("07123456789", ctx);

        _repository.VerifyAll();
        Assert.Equal(ValidationResult.Success, actual);
    }
    #endregion
}

